(GH-2) Remove vagrant from Gemfile
Vagrant is going to be deprecated as a gem in the future. This patch modifies
the instructions informing the user to download a copy of Vagrant directly
instead.

In the master release of veewee, there will be support for exporting boxes
without vagrant. However until that gem is released we still have a
documentation dependency on vagrant.

Signed-off-by: Ken Barber <ken@bob.sh>