Update README.md
Add more words around the goals of this project. I want to make it
it clear that we are not forking VeeWee templates at all, but simply
using this project as a source of documentation for rebuilds and place
to store the definitions used to build a box for the purposes of
rebuilding and so users can see what steps were followed to create the
boxes in the first place.